Move GCEFloatingIPService under GCEInternetGateway
